Moved get_assets_name and _get_assets_name to an external module to reduce code duplication.
