+++ gsy_e_sdk/setups/grid_operator_api_scripts/rest_grid_fees.py
@@ -8,7 +8,7 @@ import os
 from time import sleep
 from gsy_e_sdk.aggregator import Aggregator
 from gsy_e_sdk.rest_market import RestMarketClient
-from gsy_e_sdk.utils import log_grid_fees_information
+from gsy_e_sdk.utils import log_grid_fees_information, get_assets_name
 from gsy_e_sdk.utils import get_area_uuid_from_area_name_and_collaboration_id
 
 MARKET_NAMES = ["Grid", "Community"]
@@ -54,31 +54,6 @@ class Oracle(Aggregator):
 
     def on_finish(self, finish_info):
         self.is_finished = True
-
-
-def get_assets_name(node: dict) -> dict:
-    """
-    Parse the grid tree and return all registered assets
-    wrapper for _get_assets_name
-    """
-    if node == {}:
-        return {}
-    reg_assets = {"Area": [], "Load": [], "PV": [], "Storage": []}
-    _get_assets_name(node, reg_assets)
-    return reg_assets
-
-
-def _get_assets_name(node: dict, reg_assets: dict):
-    """
-    Parse the Collaboration / Canary Network registry
-    Return a list of the Market nodes the user is registered to
-    """
-    if node.get("registered") is True:
-        area_type = node["type"]
-        reg_assets[area_type].append(node["name"])
-    for child in node.get("children", []):
-        _get_assets_name(child, reg_assets)
-
 
 market_args = {
     "simulation_id": os.environ["API_CLIENT_SIMULATION_ID"],
